# IndianMasterApp-Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        description: Pagination metadata or other metadata
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@indianmaster.com
    name: API Support
  description: Backend APIs for Indian Master Application - A comprehensive platform
    for job seekers and hirers in India
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Indian Master API
  version: "1.0"
paths:
...
  /businesses:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of businesses filtered by business type
      parameters:
      - description: Business type to filter by
        in: query
        name: type
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g., created_at, business_name)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Businesses retrieved successfully with pagination metadata
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch businesses
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Businesses by Type
      tags:
      - Business
  /businesses/active:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all active businesses
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g., created_at, business_name)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Active businesses retrieved successfully with pagination metadata
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch active businesses
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Active Businesses
      tags:
      - Business
  /businesses/owner/{owner_id}:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of businesses owned by a specific user
      parameters:
      - description: Owner ID
        in: path
        name: owner_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g., created_at, business_name)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Businesses retrieved successfully with pagination metadata
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch businesses
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Businesses by Owner
      tags:
      - Business
...
schemes:
- http
- https
swagger: "2.0"